+++ RELEASE_NOTES.md
@@ -2,3 +2,5 @@
 
 - Update the README.md to add the installation instructions.
 - Add a new tool `isomatch tools valtable` to extract per-transcript attribute values from source GTFs and assemble them into a matrix aligned to the merged GTF transcript order.
+- support large GTF that have more than 1 billion transcripts.
+- add --purge option to only keep the transcript id and gene id attributes in the output GTF